do more

+++ notebooks/measure_pgauss.py
@@ -177,8 +177,8 @@ def get_gal_wldeblend(*, rng, data):
 def _make_obs(gal, psf, nse, rng, n=101):
     im = galsim.Convolve([gal, psf]).drawImage(
         nx=n, ny=n, scale=0.2,
-    ).array
-    psf_im = psf.drawImage(nx=n, ny=n, scale=0.2).array
+    ).array.astype(np.float32)
+    psf_im = psf.drawImage(nx=n, ny=n, scale=0.2).array.astype(np.float32)
     cen = (n-1)/2
 
     im += rng.normal(size=im.shape, scale=nse)
